pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "pageindex-rag"
version = "0.1.0"
description = "Reasoning-based RAG system powered by PageIndex"
requires-python = ">=3.10"
dependencies = [
    # PageIndex core
    "openai>=1.101.0",
    "pymupdf>=1.26.4",
    "PyPDF2>=3.0.1",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.11.0",
    "pyyaml>=6.0.2",
    # Storage
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "alembic>=1.13",
    # Vector search
    "chromadb>=0.5",
    # API
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "python-multipart>=0.0.9",
    # Utilities
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["pageindex_rag*", "pageindex*"]
